pyproject.toml
[project]
name = "uv-news-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "asgiref>=3.8.1",
    "asyncpg>=0.30.0",
    "dash>=3.0.4",
    "dash-bootstrap-components>=2.0.3",
    "dash-bootstrap-templates>=2.1.0",
    "dash-extensions>=2.0.4",
    "datasets>=3.6.0",
    "feedparser>=6.0.11",
    "httpx>=0.28.1",
    "joblib>=1.5.1",
    "python-dotenv>=1.1.0",
    "plotly>=5.0.0",  # Added explicitly
    "redis>=6.2.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
    "vadersentiment>=3.3.2",
]
